Fomating and some fixes.
Formated the jump() function, its 1 line now.
________________________________________________
Changed the playerMask on IsGrounded() since you were able to jump all
the time even when not on the ground.
________________________________________________
Fixed the huge jump the friendlyNPC was doing by increasing the height
range it follows to 4.5. Its still easily exploitable if you just while
frienlyNPC is landing.